Struct google_pubsub1::ModifyAckDeadlineRequest

pub struct ModifyAckDeadlineRequest {
    pub ack_deadline_seconds: Option<i32>,
    pub ack_ids: Option<Vec<String>>,
}

Request for the ModifyAckDeadline method.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

subscriptions modify ack deadline projects (request)

Fields

ack_deadline_seconds

The new ack deadline with respect to the time this request was sent to the Pub/Sub system. Must be >= 0. For example, if the value is 10, the new ack deadline will expire 10 seconds after the ModifyAckDeadline call was made. Specifying zero may immediately make the message available for another pull request.

ack_ids

List of acknowledgment IDs.